AzFunction/requirements.txt
azure-functions
azureml-core>=1.0.72